refactor: is_lost_childの関数定義の移動
