+++ server/src/LostChild/LostChildrenModel.py
@@ -22,3 +22,14 @@ def delete_lost(major: str, minor: str) -> None:
     :param str minor:
     """
     lost_children.delete_many({'major': major, 'minor': minor})
+
+
+def is_lost_child(major: str, minor: str) -> bool:
+    """
+    majorとminorがlost_childrenに存在するかどうか
+
+    :param str major:
+    :param str minor:
+    :return: bool
+    """
+    return bool(lost_children.find_one({'major': major, 'minor': minor}))
